+++ CHANGES.md
@@ -67,6 +67,7 @@ IronBee v0.10.0
 - Added `write_clipp` module which adds actions to write out current transaction or connection in CLIPP PB format.
 - The `libinjection` module now supports optinal confidence ratings for each pattern.  Simply add a space and the rating (0 to 100) after each line.
 - The `@is_sqli` operator (`libinjection` module) now supports capture.  The matching fingerprint will be stored under the `fingerprint` key and the confidence under the `confidence` key.
+- Renamed the `libinjection` module directive `SQLiPatternSet` (now deprecated) to `LibInjectionFingerprintSet` to better match libinjection terminology and directive naming conventions.
 
 **Predicate**